Created and coded HookGrapple and ScissorLift
HookGrapple is the cascading slides that will lift the robot off the ground and onto the lander. ScissorLift is a mechanism that may be used for lifting gold and silver pieces to their designated boxes.

+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Drivetrains/JacobDrive.java
@@ -16,7 +16,7 @@ public class JacobDrive {
         this.leftMotor = left;
         this.rightMotor = right;
     }
-    public void jDrive(Float power){
+    public void jDrive(double power){
         rightMotor.setPower(power);
         leftMotor.setPower(-power);
     }
